restapi/requirements.txt
argon2==0.1.10
authorizenet==1.1.4
Flask==3.0.3
Flask_Cors==4.0.0
Flask_RESTful==0.3.10
fpdf2==2.7.8
pdf2image==1.17.0
Pillow==10.3.0
pycryptodome==3.20.0
PyJWT==2.8.0
PyJWT==2.8.0
PyMySQL==1.1.0
PyPDF2==3.0.1
pyperclip==1.8.2
pytz==2023.3.post1
Requests==2.32.2
